src/test/kotlin/org/example/e3FxGaming/adventOfCode2022Kotlin/Day04Test.kt
package org.example.e3FxGaming.adventOfCode2022Kotlin

import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.params.ParameterizedTest
import org.junit.jupiter.params.provider.Arguments
import org.junit.jupiter.params.provider.MethodSource
import java.io.Reader
import java.util.stream.Stream

internal class Day04Test {
    companion object {
        @JvmStatic
        fun argumentsProvider(): Stream<Arguments> = Stream.of(
            Arguments.of(Day04Test::class.java.getResource("/input/day04.txt").toReader(), 2, 4)
        )
    }

    @ParameterizedTest
    @MethodSource("argumentsProvider")
    fun day04(inputReader: Reader, part1: Int, part2: Int) {
        Day04(inputReader).let {
            assertEquals(part1, it.part1(), "Part 1")
            assertEquals(part2, it.part2(), "Part 2")
        }
    }
}